# nao_backend/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:5001",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    },
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "https://petstore.swagger.io/oauth/authorize",
      "flow": "implicit",
      "scopes": {
        "read_pets": "read your pets",
        "write_pets": "modify pets in your account"
      }
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/cards/groupsId/{id}": {
      "get": {
        "tags": [
          "Cards"
        ],
        "description": "Get all cards of a group",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/Cards"
            },
            "description": "Get all cards of a group"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "No cards found"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "No cards with this groupsId exist"
          }
        }
      }
    },
    "/cards/userId": {
      "post": {
        "tags": [
          "Cards"
        ],
        "description": "Create card",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "groupsId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "Created new Card"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Created card"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Card already exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Card already exist"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error creating card"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error creating card"
          }
        }
      }
    },
    "/cards/delete/{name}": {
      "delete": {
        "tags": [
          "Cards"
        ],
        "description": "Delete card",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "Deleted card"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Deleted card"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Card doesnt exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Card doesn't exists"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error deleting card"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error deleting card"
          }
        }
      }
    },
...
    "/groups/userId/{id}": {
      "get": {
        "tags": [
          "Groups"
        ],
        "description": "Get all groups of an organization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/Groups"
            },
            "description": "Get organization by name"
          },
          "203": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "No groups found"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "No groups found"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Organization dosn't exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Organization dosn't exists"
          }
        }
      }
    },
    "/groups/userId/": {
      "post": {
        "tags": [
          "Groups"
        ],
        "description": "Create group in organization",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "organizationId": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "$ref": "#/definitions/CreateGroup"
            },
            "description": "Created new Group"
          },
          "405": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Group already exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Group already exists"
          },
          "409": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Organization dosn't exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Organization dosn't exist"
          },
          "500": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Error creating Group"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Error creating Group"
          }
        }
      }
    },
    "/groups/delete/{name}": {
      "delete": {
        "tags": [
          "Groups"
        ],
        "description": "Delete group by name",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "schema": {
              "type": "object",
              "properties": {
                "result": {
                  "type": "string",
                  "example": "Deleted group"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Deleted group"
          },
          "404": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Group dosn't exists"
                }
              },
              "xml": {
                "name": "main"
              }
            },
            "description": "Group dosn't exists"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Cards": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 15
              },
              "name": {
                "type": "string",
                "example": "cat"
              },
              "groupsId": {
                "type": "number",
                "example": 13
              }
            }
          }
        }
      }
    },
...
    "Groups": {
      "type": "object",
      "properties": {
        "result": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "id": {
                "type": "number",
                "example": 36
              },
              "name": {
                "type": "string",
                "example": "food"
              },
              "organizationName": {
                "type": "string",
                "example": "Epitech"
              }
            }
          }
        }
      }
    },
    "CreateGroup": {
      "type": "object",
      "properties": {
        "groupeName": {
          "type": "string",
          "example": "Jean"
        },
        "organizationId": {
          "type": "number",
          "example": 18
        }
      }
    },
...
  }
}